Renamed VOMS2CSSyncronizer -> VOMS2CSSynchronizer
